% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{create_grid}
\alias{create_grid}
\title{Creat A Grid}
\usage{
create_grid(grid_size, n_evidence)
}
\arguments{
\item{grid_size}{a numeric vector of the form c(nrows, ncolumns)
which gives the size of the grid in x and y direction.}

\item{n_evidence}{a number specifing the amount of evidence in the grid.}
}
\value{
A matrix containing a grid.
}
\description{
\code{create_grid} creates a grid from the given set of values.
}
\details{
The user can specify how big the grid is and how much evidence will
be placed in the grid. The function first creates a grid that is completely
empty. The evidence is sampled and added to the grid at random. In a last step
a wall is being built around the grid.
}
\examples{
create_grid(grid_size = c(10, 10), n_evidence = 5)
}
