% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{create_population}
\alias{create_population}
\title{Creating A Population}
\usage{
create_population(individuals)
}
\arguments{
\item{individuals}{a number}
}
\description{
\code{create_poulation} creates a population of x individual solutions.
The robot can look at 5 sites: the current field and the fields north, east,
south, and west of him. Each of the sites can contain a wall, evidence or nothing.
Following, the robot can be in 243 possible situations.
For each individual strategy table is created: 243 moves are randomly sampled
and assigned to each of the situations.
}
\examples{
create_population(100)
}
