% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{life_cycle}
\alias{life_cycle}
\title{One Life Cycle of a Population}
\usage{
life_cycle(population, grid_size, n_evidence, steps, sessions)
}
\arguments{
\item{population}{a list containing a population made with \code{create_population}.}

\item{grid_size}{a numeric vector of the form c(nrow, ncol)
which gives the size of the grid.}

\item{n_evidence}{a number specifing the amount of evidence in the grid.}

\item{steps}{a number indicating how many moves the robot should walk in one
grid configuration.}

\item{sessions}{a number indicating how many times the grid configuration is
changed.}
}
\value{
The score for each individual strategy in each session.
}
\description{
\code{life_cycle} is used to simulate one life cycle of a population.
}
\details{
The fitness of an individual strategy is determined by seeing how well
the strategy works in X different sessions.

In one session, the robot walks X steps in a grid and is scored on his actions
(see function \code{move_score}).
A new session is initiated with changing the configuration of the grid. Consequently,
the robot walks again and is scored. The number of sessions is indicated by
the user with the argument \code{sessions}.

This procedure is done for every individual strategy in the population.
}
\examples{
first_population <- create_population(50)
life_cycle(population = first_population, grid_size = c(10, 10), n_evidence = 10,
steps = 100, sessions = 200)
}
