% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{lookup_situation}
\alias{lookup_situation}
\title{Retrieve the Number of the Current Situation}
\usage{
lookup_situation(grid, latitude = 2, longitude = 2)
}
\arguments{
\item{grid}{a matrix.}

\item{latitude}{a number indicating the current position of the robot on the y-axis.}

\item{longitude}{a number indicating the current position of the robot on the xaxis.}
}
\description{
\code{lookup_situation} looks up the robots current situation in his strategy table.
}
\details{
From wherever the robot is, he can see the content of one adjacent site
in the north, south, east, and west, as well as the content of the site he occupies.
To decide what to do next, the robot looks up his current situation in his strategytable.
There he finds the corresponding action.
}
\examples{
lookup_situation(grid, latitude = 9, longitude = 5)
}
