% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualize_path}
\alias{visualize_path}
\title{Visualization of the Path}
\usage{
visualize_path(individual, grid_size, evidence_latitude, evidence_longitude,
steps, animal)
}
\arguments{
\item{individual}{one individual strategy taken form the population made with
\code{create_poulation}.}

\item{grid_size}{a numeric vector of the form c(nrows, ncolumns)
which gives the size of the grid in x and y direction.}

\item{evidence_latitude}{a string indicating the position of the evidence on
the x-axis.}

\item{evidence_longitude}{a string indicating the position of the evidence
on the y-axis.}

\item{steps}{a number indicating how many moves the robot should walk in the
grid.}

\item{animal}{the animal that is being searched (see details).}
}
\value{
an animated plot.
}
\description{
\code{visualize_path} is used to plot the robot's moves.
}
\details{
The robot begins his path at his starting position (latitude = 2, longitude = 2).
The robot then follows the given strategy \code{individual} for X \code{steps}.
If there is no wall in the direction of his next movement, the robot walks.
If he crashes into a wall, he bounces back into the
current site.
If his next action is to stay or to pick up, the robot stops moving.
The robot's path is visualized with \code{ggplot}.

#' \code{visualize_path} can only display the following animal footprints: cat, horse,
bird or bear.
}
\examples{
population1 <- create_population(30)
visualize_path(individual = population1[[3]], grid_size = c(5, 7),
evidence_latitude = c(3, 2, 4), evidence_longitude = c(1, 6, 5),
steps = 30, animal = "bird")
}
