% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{create_grid}
\alias{create_grid}
\title{Creat A Grid}
\usage{
create_grid(grid_size, n_evidence)
}
\arguments{
\item{grid_size}{a numeric vector of the form c(max of x-axis, max of y-axis)
which gives the size of the grid in x and y direction.}

\item{n_evidence}{a numeric specifing the number of evidence in the grid.}
}
\value{
a matrix containing a grid
}
\description{
\code{create_grid} creates a grid. The user can specify how big the grid is and
how much evidence will be placed in the grid. There is a wall around the grid
(first row, first column, last row, last column)
}
\examples{
create_grid(grid_size = c(10, 10), n_evidence = 5)
}
