% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{create_population}
\alias{create_population}
\title{Create A Population}
\usage{
create_population(individuals)
}
\arguments{
\item{individuals}{a number indicating how many individuals are in one population.}
}
\value{
A list containing a population of x data.frames of solutions.
}
\description{
\code{create_poulation} A function to create the primary population of
candidate strategies.
}
\details{
There are five sites the robot can see: the current field and the
fields north, east, south, and west of the robot. Each of the sites can
contain a wall, evidence or nothing. Consequently, the robot can possibly be
in 243 situations. In each situation, the robot can perform one of the following
six actions: move north, east, south, and west, stay in one field, or pick up
something. An individual is one specific strategy to move through the grid.
\code{create_poulation} creates for each individual in the population a random
strategy table:
243 moves are randomly sampled and assigned to each of the situations.
}
\examples{
create_population(100)
}
