% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{lookup_situation}
\alias{lookup_situation}
\title{Retrieve from the Strategy Table the Number of the Current Situation}
\usage{
lookup_situation(individual, grid, latitude = 2, longitude = 2)
}
\arguments{
\item{individual}{one individual strategy taken form the population made with
\code{create_poulation}.}

\item{grid}{a matrix made with \code{create_grid}.}

\item{latitude}{a number indicating the current position of the robot on the
y-axis.}

\item{longitude}{a number indicating the current position of the robot on the
x-axis.}
}
\value{
A number from the robots' strategy table.
}
\description{
\code{lookup_situation} retrieves the robots current situation in his strategy
table.
}
\details{
To decide which move to perform next, the robot looks up his current
situation in his strategy table. There he finds the corresponding action.
}
\examples{
population1 <- create_population(50)
my_grid <- create_grid(c(5, 5), 8)
lookup_situation(individual = population1[[2]], grid = my_grid, latitude = 9,
longitude = 5)
}
